airinsights-backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
python-dateutil==2.8.2
